requirements.txt
aiogram==3.4.1
asyncpg
sqlalchemy
python-dotenv